refactor(random): move to crypto folder
