+++ lib/crypto/random.go
@@ -1,4 +1,4 @@
-package tutil
+package crypto
 
 import (
 	"crypto/rand"
